requirements.txt
streamlit
pandas
requests
joblib
scikit-learn
plotly
python-dotenv
streamlit-autofresh
numpy